+++ subscription/mongodb/native/blocking/src/main/java/org/occurrent/subscription/mongodb/nativedriver/blocking/NativeMongoSubscriptionModel.java
@@ -17,6 +17,7 @@
 package org.occurrent.subscription.mongodb.nativedriver.blocking;
 
 import com.mongodb.MongoClientSettings;
+import com.mongodb.MongoCommandException;
 import com.mongodb.MongoException;
 import com.mongodb.client.ChangeStreamIterable;
 import com.mongodb.client.MongoChangeStreamCursor;
@@ -49,6 +50,7 @@ import org.slf4j.LoggerFactory;
 import javax.annotation.PreDestroy;
 import java.time.Duration;
 import java.util.Collections;
+import java.util.Date;
 import java.util.List;
 import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
@@ -65,9 +67,8 @@ import static com.mongodb.client.model.Aggregates.match;
 import static java.util.Objects.requireNonNull;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.SECONDS;
-import static org.occurrent.retry.Backoff.exponential;
-import static org.occurrent.retry.RetryStrategy.retry;
 import static org.occurrent.retry.internal.RetryExecution.executeWithRetry;
+import static org.occurrent.subscription.mongodb.internal.MongoCommons.bsonTimestampNow;
 
 /**
  * This is a subscription that uses the "native" MongoDB Java driver (sync) to listen to changes from the event store.
@@ -263,7 +264,14 @@ public class NativeMongoSubscriptionModel implements PositionAwareSubscriptionMo
     public SubscriptionPosition globalSubscriptionPosition() {
         // Note that we increase the "increment" by 1 in order to not clash with an existing event in the event store.
         // This is so that we can avoid duplicates in certain rare cases when replaying events.
-        BsonTimestamp currentOperationTime = MongoCommons.getServerOperationTime(database.runCommand(new Document("hostInfo", 1)), 1);
+        BsonTimestamp currentOperationTime = null;
+        try {
+             currentOperationTime = MongoCommons.getServerOperationTime(database.runCommand(new Document("hostInfo", 1)), 1);
+        } catch (MongoCommandException e) {
+            // This can if the server doesn't allow to get the operation time since "db.adminCommand( { "hostInfo" : 1 } )" is prohibited.
+            // This is the case on for example shared Atlas clusters. If this happens we return the current time of the client instead.
+            currentOperationTime = bsonTimestampNow();
+        }
         return new MongoOperationTimeSubscriptionPosition(currentOperationTime);
     }